:art: result-box 관련 함수들 모아서 모듈 생성
